+++ appinfo/app.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * ownCloud - eudat
+ * ownCloud - b2sharebridge
  *
  * This file is licensed under the Affero General Public License version 3 or
  * later. See the LICENSE file.
@@ -9,7 +9,7 @@
  * @copyright EUDAT 2015
  */
 
-namespace OCA\Eudat\AppInfo;
+namespace OCA\B2shareBridge\AppInfo;
 
 use OCP\App;
 use OCP\Util;
@@ -22,12 +22,12 @@ $navigationEntry = function () use ($c) {
         'id' => $c->getAppName(),
         'order' => 100,
         'name' => $c->query('EudatL10N')->t('B2SHARE'),
-        'href' => $c->getServer()->getURLGenerator()->linkToRoute('eudat.Eudat.index'),
-        'icon' => $c->getServer()->getURLGenerator()->imagePath('eudat', 'appbrowsericon.svg'),
+        'href' => $c->getServer()->getURLGenerator()->linkToRoute('b2sharebridge.B2shareBridge.index'),
+        'icon' => $c->getServer()->getURLGenerator()->imagePath('b2sharebridge', 'appbrowsericon.svg'),
     ];
 };
 
 $c->getServer()->getNavigationManager()->add($navigationEntry);
 
-App::registerAdmin('eudat', 'lib/settings');
-Util::addScript('eudat', 'fileactions');
+App::registerAdmin('b2sharebridge', 'lib/settings');
+Util::addScript('b2sharebridge', 'fileactions');