+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 eudat (0.0.1)
 * First release
+
+eudat (0.0.4)
+* version with working third-party transfer
+
+eudat (0.0.5)
+* rename of module from eudat to b2sharebridge